Lang: Implement full datatype completion
Datatype completions means that we recurse through the whole AST and
ensure that every IdentifierNode has its correct type. This is important
for the later stages of compilation.

Example:
x = fib(32);

The code will now try to find a fib function, figure out its return type
and set x to that type.