Lang: VM: Compile whole program to a single chunk of jit-assembly
Previously, we split the program into individual chunks which were
either JIT-chunks with a program or interpret-chunks. Now we instead
create one giant jit-program that just exists anytime it doesn't know
how to handle an instruction.

This increases performance in workloads were not every instruction can
be jitted by about 3 to 4 times.