Lang: VM: Replace weird jump list with a real jump table
I finally got it to work after a lot of time and it works great, but
only makes a difference for large programs; if the program is small, it
can even be slower.

To test this out, I added 6 copies of the fib function before the one
that I'm actually calling. These copies cause the actual fib function to
have larger instruction pointers meaning that our old code had to do
more comparisions; for the jump table, it doesn't matter how large the
code is, it's always constant.

In this test the new jump table was about 2.3 times faster than the
comparision list!

To further improve this, we should replace the return address (currently
0x42424242) with a value that we can jump to directly, probably 0x0. If
this instruction is returned to, we will exit the VM. This also means
that every other instruction will need to be shifted by one. This will
remove the need for a weird comparision after every return to check if
we should exit.