Lang: Implement tuple access syntax (tuple:1) & fix vm for it
The syntax uses a colon instead of a dot because the dot gets parsed as
the identifier name right now - this should be fixed in the parser soon.

This patch also fixes the REPUSH_FROM_N instruction in JIT mode, which
only caused a problem when you repush more than 8 bytes at once (which
is right now necessary for tuple access).

TODO don't repush the whole tuple if it's referred by it's name and
instead fetch the accessed element directly from where it is stored on
the stack.