Lang: Fix jit performance regression introduced by b3bf37ff10
This code is even about 30% faster than before b3bf37ff10 in the fib(28)
benchmark.

I noticed that performance was about twice as bad after b3bf37ff10 as it
had been before that patch, but that patch only introduced 1 x86
instruction for 3 of the bytecode instructions. The real reason for the
large regression seems to be the fact that movsq() is faster with
the direction flag (std()) set than without it (cld()).

This patch replaces the code that previously needed a cleared direction
flag (cld()) with one that works without movsq and only uses mov and
push.