Lang: WIP: Add template datatype completion
Whenever we now encounter an identifier like add<i32> during datatype
completion, we check if the identifier add was already used with the
template parameter <i32>. If not, we add it to list and return its
index, if it was already added, then we just add it to the list first.

This index is stored in the IdentifierNode of the add<i32> function call
to be later used during compilation to distinguish different template
instantiations like add<i32> and add<i64>.

The DatatypeCompleter returns a list of all instantiated templates to be
later used by the compiler to compile each one.

The code is quite ugly right now but it seems do its job. This will
probably require a few rewrites to get it to look nice.