Lang: Fix fundamental build error leading to SEGFAULT when JIT enabled
This is a bug in a league of its own.

When compiling the library, we correctly set the define-values
depending on if you want to compile with jit or without. This makes
a difference to the VM.hpp header. The problem was that we didn't
set the same flags in the tests & cli program, meaning when those files
included VM.hpp, they got a slightly different version with one member
misssing, leading to a crash in the destructor.

This was fixed by setting the same flags in samal_cli and tests, but in
the future we should maybe have a common config.hpp so that user code
doesn't need to define the same flags.

Please clang/gcc gods, give us C++ modules! (Idk if those would have
helped, but that would still have been nice. Also, a compiler warning
would have been nice.)