Lang: Save information about stack in Program & use it to dump vars in the VM
This patch adds the StackInformationTree datastructure which stores
information about variables on the stack and is generated at
compile-time.

This is useful for a potential debugger and - more importantly - for a
GC. This is because the GC needs to know that datatype of each variable
on stack to correctly recurse through potential pointers etc.