Lang: Fix undefined behaviour in jit entrance/exit code
TIL that there is something called a "red zone" on linux which is the
128 bytes below rsp; if you try to access anything below that, it's not
allowed. This means that when the stack is larger than 128 bytes and we
enter or exit the VM, we previously wrote beyond this safe red zone.

To fix this rsp is now moved first. I thought accessing any memory on
the stack was safe...