Lang: Fix jit <-> interpreter incompatibilities and improve switch speed
Since the interpreter now works with a reversed stack, there is no need
anymore to reverse the stack everytime before and after we execute jit
code. This improves performance in a small list-heavy benchmark by about
2 times.

This patch also includes a number of bug fixes to make all this possible.