Lang: Improve jit <-> interpreter switch performance & general jit speed
This patch makes my list benchmark run about 20 times faster, now
beating elixir/BEAM by a factor of 2 if you include the dump at the end
(though the BEAM doesn't jit). If you don't include the dump, the speed
is about the same. That's a great success!

This was mostly done by removing the copying of the interpreter-stack
to/from the cpu stack every time we enter/exit the jit environment.
Instead, we now operate on the interpreter-stack directly. I was quiete
suprised that the CPU/Linux doesn't really care where my rsp points to,
though I imagine it could maybe cause problems with signals?

In addition to that, I improved the speed of the RETURN and the
POP_N_BELOW instructions, mostly because the old version didn't work
with the new system anymore for some reason. This alone makes the fib
benchmark about 25% faster.

That said, I imagine there is still quiete a lot of room for
optimizations like merging POP_N_BELOW instructions and much more.