Lang: WIP Implement very inefficient GC
The function VM::generateStacktrace still doesn't work completely fine :c,
TODO fix this function once and for all.

The GC I implemented is VERY inefficient, probably because it allocates
as well (ironic, he could save other, but not himself). Because of this
the GC needs to run quiete rarely if you want any speed at all. TODO
optimize this.