Lang+Code+Tests: Fully implement enum creation support
This patch also fixes weird behaviour that struct creation parameters
where executed in reverse order - this has now been fixed.

Enums are allocated on the heap to allow recursive enums, similar to
structs. I also added tests to ensure that everything works.

I recognized while writing this code that structs actually
don't need to be heap allocated anymore because the only way to
construct a recursive struct anyways is to use enums or lists, both of
which are heap allocated. This is because a struct can't actually be
null which makes constructing simple recursive structs impossible.