Lang: Put struct data now on stack instead of the heap
This is part of a bigger project of adding a new pointer datatype that
can point to any contained type, similar to a list. This means we don't
need to put every struct/enum on the heap anymore to allow recursive
datatypes. As we now have the choice I'm moving everything to the stack
because it's way faster because our GC is horribly slow.