Lang: Fix Datatype::operator== not checking equality of partial types correctly
This is a bug that was just introduced last patch or so where we started
checking if two structs have the same name and, if they do, saying
they're equal - disregarding the fact that they could have different
parameter types. Now we have a special case for undetermined identifiers
which completes them first and the compares the results.