Lang+Tests: Correctly check equality of template types (+ add test)
Previously, we only checked if the names (strings) of two undetermined
identifiers are the same; this doesn't work, because two T's can refer
to different types. Always resolving these types doesn't work either,
because a struct/enum type could be recursive.

We now track wether a give undetermined identifier replacement map entry
was created by a user type or by a template parameter and only check
types recursively if the type is a template parameter.