Lang+Code: Fix not completing user-types across modules correctly
Previously, we only considered the module we're calling a function from
for full type deduction (like matching Vec2 to Structs.Vec2). This of
course doesn't work if we're in a different module, like calling a
function from the module Structs that returns a Vec2 from the module
Main.

We now always consider the usingModuleNames of the module where a given
function is from.

TODO cleanup Compiler::compileIdentifierLoad code, it's a mess.