Lang: Make GC::markAndSweep *way* more efficient
Previously, we used a simple mark-and-sweep garbage collector. This
meant that in list-heavy task the gc pressure was insane, leading to
almost 50% of the execution time being spent in the GC.

This patch replaces the mark-and-sweep algorithm with a copying garbage
collector, which is very efficient if you allocate a lot of mememory but
don't actually use most of it. The implementation is quite naive right
now and doesn't make use of fancy multi-region ephermal/generational
garbage collection, but it's still way better than the old one.

Garbage collection now only takes about 1-20µs in the euler benchmark on
my pc, making it almost free. Allocation is even cheaper, almost free now
as well.