Lang: Add functions for instantiating enums from native code
This patch adds a function called Pipeline::type (and
Pipeline::incompleteType) which allows you to create Datatype objects
from strings directly. This makes writing native functions way easier,
but is quiete slow, as it utilizes the peg parser in the background.
It's recommended to only use this function during startup and store the
returned types somewhere for later use.