Lang+Tests: Fix complex GC bug freeing memory that is still used
The bug appeared in code like this: ["Hello", Core.identity("Welt")]. If
the GC now runs in Core.identity by chance then "Hello" will be freed.
This leads to UB if the variable is later used. This could happen with
function calls and tuples as well.

This was fixed by effectively compiling the above code as:
{ [param$1 = "Hello", param$2 = Core.identity("Welt")] } (though is is
strictly speaking not valid samal code as assignments are statements,
not expressions).

We use the new tinyStackFrame system to store references to the
parameters to stop them from being removed from the GC. The only problem
was the compileTupleCreation functions, as tuples don't have a magic
CREATE_TUPLE instruction. Due to the weirdniss of saving stack frames
related to instructions, we had to add the NOOP instruction for this
case.

This patch also adds tests to ensure that these bugs stay fixed.