Lang: Fix bug in Compiler::compileIdentifierLoad
This bug meant that the function called addLabel but then threw an
exception because it couldn't load infer the function type. This was
then caught by the function call compiler which tried to infer the
arguemnts and then called addLabel himself, effectively dropping the
first label. We now only add a label in compileIdentifierLoad if we
could infer the function type.