Lang: Fix bug in cross-module typename resolution
This bug showed up when you did the following:
Module A:
struct a {
	member : b
}
struct b {
	val : i32
}

Module B:
fn main() -> () {
	a = A.a{member: A.b{val : 5}}
}

When trying to get the size of A.a in module B, we need to get the type
of each member/field. For that we use the current module-using scope,
but because the struct is defined in another module, we won't find A.b.

To fix this, each value in the UndeterminedIdentifierReplacementMap now
contains a field that lists all the module names that are 'using'-ed in
the module where that variable comes from.